Allow slight mismatch in weight
Mostly copied from xst[0], to avoid "font weight does not match" errors
and using default attributes when the font weight metadata is a bit off.
I've particularly seen this happen with the Go Mono font, and previously
worked around it by making a copy of Go Mono Regular with a distinct
name so that the other Go Mono weights don't match, although now that
I'm looking at the code I'm surprised that worked.

[0]: https://github.com/gnotclub/xst/pull/52/commits/e81286c21c50fa2d6e2437eb9017de009f642611